# IrumOrder_BackEnd/src/main/resources/docs/swagger-order.yaml
...
openapi: 3.0.1
info:
  title: API - 이룸오더
  description: API Description
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  "/orders/order":
    post:
      tags:
        - Order Management
      summary: Create Order
      description: Creates a new order based on provided details.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OrderRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/OrderResponseDto"
  "/orders/orderCheck":
    post:
      tags:
        - Order Management
      operationId: getAllOrdersByUserId
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/OrdersCheckResponseDto"
...
components:
  schemas:
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
      description: 루틴이 실행될 시간
      example: '09:00'
...
    MenuDetailDto:
      type: object
      properties:
        useCup:
          type: string
          example: TakeOut
        addShot:
          type: boolean
          example: true
        addVanilla:
          type: boolean
          example: false
        addHazelnut:
          type: boolean
          example: true
        light:
          type: boolean
          example: true
    OrderMenuDto:
      type: object
      properties:
        menuId:
          type: integer
          format: int32
          example: 1
        quantity:
          type: integer
          format: int32
          example: 2
        menuOptions:
          "$ref": "#/components/schemas/MenuDetailDto"
    OrderRequestDto:
      type: object
      properties:
        userId:
          type: integer
          format: int32
          example: 1
        totalPrice:
          type: integer
          format: int32
          example: 3200
        pickUp:
          "$ref": "#/components/schemas/LocalTime"
        orderMenuOptions:
          type: array
          items:
            "$ref": "#/components/schemas/OrderMenuDto"
    OrderResponseDto:
      type: object
      properties:
        orderId:
          type: integer
          format: int32
          example: 1
        userId:
          type: integer
          format: int32
          example: 1
        totalPrice:
          type: integer
          format: int32
          example: 3000
        orderStatus:
          type: string
          example: WAITING
          enum:
            - NEW_ORDER
            - WAITING
            - ACCEPTED_ORDER
            - READY_FOR_PICKUP
            - COMPLETED
            - ACCEPTED
            - COOKING_START
            - COOKING_END
            - PICKUP_END
        pickUp:
          "$ref": "#/components/schemas/LocalTime"
        payment:
          type: string
          format: date-time
        orderMenuOptions:
          type: array
          items:
            "$ref": "#/components/schemas/OrderMenuDto"
    OrderDto:
      type: object
      properties:
        menuId:
          type: integer
          format: int32
          example: 1
        menuName:
          type: string
          example: Americano
        quantity:
          type: integer
          format: int32
          example: 2
        useCup:
          type: string
          example: 'true'
        addShot:
          type: boolean
          example: false
        addVanilla:
          type: boolean
          example: true
        addHazelnut:
          type: boolean
          example: false
        light:
          type: boolean
          example: true
    OrdersCheckResponseDto:
      type: object
      properties:
        orderId:
          type: integer
          format: int32
          example: 1
        userId:
          type: integer
          format: int32
          example: 1
        totalPrice:
          type: integer
          format: int32
          example: 3000
        orderStatus:
          type: string
          example: WAITING
          enum:
            - NEW_ORDER
            - WAITING
            - ACCEPTED_ORDER
            - READY_FOR_PICKUP
            - COMPLETED
            - ACCEPTED
            - COOKING_START
            - COOKING_END
            - PICKUP_END
        pickUp:
          "$ref": "#/components/schemas/LocalTime"
        orderMenuOptions:
          type: array
          items:
            "$ref": "#/components/schemas/OrderDto"
...